Fix cursor jumping bug by wrapping Transition Note textarea
